% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_cor_function.R
\name{year_cor}
\alias{year_cor}
\title{A Year Correlation Function}
\usage{
year_cor(year)
}
\arguments{
\item{year}{an individual year}
}
\value{
matrix
}
\description{
This function allows you to get the corrlation matrix for all numeric variables in the data for that year.
}
\examples{
year_cor(2013)

}
