% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_oldest_function.R
\name{year_oldest}
\alias{year_oldest}
\title{A Year Age Function}
\usage{
year_oldest(year)
}
\arguments{
\item{year}{an individual year}
}
\value{
character
}
\description{
This function allows you to find the name of the player who was the odest player in NBA data for that year.
}
\examples{
year_oldest(2013)

}
