% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_point_function.R
\name{year_point}
\alias{year_point}
\title{A Year Point Function}
\usage{
year_point(year)
}
\arguments{
\item{year}{an individual year}
}
\value{
character
}
\description{
This function allows you to find the name of the player who got highest points for that year.
}
\examples{
year_point(2013)

}
